requirements.txt
fastapi>=0.104.0,<0.110.0
uvicorn[standard]>=0.24.0,<0.28.0
pydantic>=2.5.0,<3.0.0
anthropic>=0.18.0,<0.20.0
openai>=1.12.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.25.0,<0.27.0
azure-ai-textanalytics>=5.3.0,<6.0.0
azure-core>=1.30.0,<2.0.0
azure-ai-documentintelligence>=1.0.0,<2.0.0
azure-cognitiveservices-speech>=1.40.0,<2.0.0
python-multipart>=0.0.9,<0.1.0
supabase>=2.0.0,<3.0.0
